platform: tegra: Fix overflow before widen
To avoid overflow, cast mc_all_actives to type u64.

Bug 3952896

CID 10174209

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nv-oot/+/2930556
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>

+++ drivers/platform/tegra/tegra-cactmon-mc-all.c
@@ -75,7 +75,7 @@ static int central_actmon_mc_all_get(void *data, u64 *val)
 
 	mc_all_actives = cactmon_readl(cactmon, CENTRAL_ACTMON_MC_ALL_AVG_COUNT_REG);
 
-	*val = mc_all_actives * 1000 / sample_period_usec;
+	*val = (u64)mc_all_actives * 1000 / sample_period_usec;
 
 	return 0;
 }
